AWS-Textract/swedish_pdf_processor/requirements.txt
boto3>=1.26.0
pdf2image>=1.16.0
pillow>=9.0.0
tqdm>=4.64.0
pandas>=1.5.0
openpyxl>=3.0.10
pytest>=7.0.0
pdfplumber>=0.7.0
# Notes:
# poppler-utils is a system dependency for pdf2image
# Install via: apt-get install poppler-utils (Ubuntu/Debian) 
# or brew install poppler (macOS)
# For Windows, download from: https://github.com/oschwartz10612/poppler-windows/releases